fix failing test and refactor others into appropriate suites
