+++ tests/src/test/scala/system/health/AlarmsHealthFeedTests.scala
@@ -18,10 +18,10 @@ package system.health
 
 import common._
 import org.junit.runner.RunWith
+import org.scalatest.FlatSpec
 import org.scalatest.junit.JUnitRunner
-import org.scalatest.{FlatSpec, Inside}
-import spray.json.DefaultJsonProtocol.{BooleanJsonFormat, IntJsonFormat, LongJsonFormat, StringJsonFormat}
-import spray.json.{JsObject, JsString, pimpAny}
+import spray.json.DefaultJsonProtocol.{IntJsonFormat, LongJsonFormat, StringJsonFormat}
+import spray.json.pimpAny
 
 /**
  * Tests for alarms trigger service
@@ -30,7 +30,6 @@ import spray.json.{JsObject, JsString, pimpAny}
 class AlarmsHealthFeedTests
     extends FlatSpec
     with TestHelpers
-    with Inside
     with WskTestHelpers {
 
     val wskprops = WskProps()
@@ -196,274 +195,4 @@ class AlarmsHealthFeedTests
             activationsAfterInterval should be(2)
     }
 
-    it should "update cron, startDate and stopDate parameters" in withAssetCleaner(wskprops) {
-        (wp, assetHelper) =>
-            implicit val wskProps = wp
-            val triggerName = s"dummyAlarmsTrigger-${System.currentTimeMillis}"
-            val packageName = "dummyAlarmsPackage"
-
-            // the package alarms should be there
-            val packageGetResult = wsk.pkg.get("/whisk.system/alarms")
-            println("fetched package alarms")
-            packageGetResult.stdout should include("ok")
-
-            // create package binding
-            assetHelper.withCleaner(wsk.pkg, packageName) {
-                (pkg, name) => pkg.bind("/whisk.system/alarms", name)
-            }
-
-            val cron = "* * * * *"
-            val startDate = System.currentTimeMillis + (1000 * 30)
-            val stopDate = startDate + (1000 * 100)
-
-            // create trigger feed
-            println(s"Creating trigger: $triggerName")
-            assetHelper.withCleaner(wsk.trigger, triggerName) {
-                (trigger, name) =>
-                    trigger.create(name, feed = Some(s"$packageName/alarm"), parameters = Map(
-                        "cron" -> cron.toJson,
-                        "startDate" -> startDate.toJson,
-                        "stopDate" -> stopDate.toJson))
-            }
-
-
-            val actionName = s"$packageName/alarm"
-            val readRunResult = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "READ".toJson,
-                "authKey" -> wskProps.authKey.toJson
-            ))
-
-            withActivation(wsk.activation, readRunResult) {
-                activation => activation.response.success shouldBe true
-
-                    inside(activation.response.result) {
-                        case Some(result) =>
-                            val config = result.getFields("config").head.asInstanceOf[JsObject].fields
-                            val status = result.getFields("status").head.asInstanceOf[JsObject].fields
-
-                            config should contain("cron" -> cron.toJson)
-                            config should contain("startDate" -> startDate.toJson)
-                            config should contain("stopDate" -> stopDate.toJson)
-
-                            status should contain("active" -> true.toJson)
-                            status should contain key "dateChanged"
-                            status should contain key "dateChangedISO"
-                            status should not(contain key "reason")
-                    }
-            }
-
-            val updatedCron = "*/2 * * * *"
-            val updatedStartDate = System.currentTimeMillis + (1000 * 30)
-            val updatedStopDate = updatedStartDate + (1000 * 100)
-
-            val updateRunAction = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "UPDATE".toJson,
-                "authKey" -> wskProps.authKey.toJson,
-                "cron" -> updatedCron.toJson,
-                "startDate" -> updatedStartDate.toJson,
-                "stopDate" -> updatedStopDate.toJson
-            ))
-
-            withActivation(wsk.activation, updateRunAction) {
-                activation => activation.response.success shouldBe true
-            }
-
-            val runResult = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "READ".toJson,
-                "authKey" -> wskProps.authKey.toJson
-            ))
-
-            withActivation(wsk.activation, runResult) {
-                activation => activation.response.success shouldBe true
-
-                    inside(activation.response.result) {
-                        case Some(result) =>
-                            val config = result.getFields("config").head.asInstanceOf[JsObject].fields
-
-                            config should contain("cron" -> updatedCron.toJson)
-                            config should contain("startDate" -> updatedStartDate.toJson)
-                            config should contain("stopDate" -> updatedStopDate.toJson)
-                    }
-            }
-    }
-
-    it should "update fireOnce and payload parameters" in withAssetCleaner(wskprops) {
-        (wp, assetHelper) =>
-            implicit val wskProps = wp
-            val triggerName = s"dummyAlarmsTrigger-${System.currentTimeMillis}"
-            val packageName = "dummyAlarmsPackage"
-
-            // the package alarms should be there
-            val packageGetResult = wsk.pkg.get("/whisk.system/alarms")
-            println("fetched package alarms")
-            packageGetResult.stdout should include("ok")
-
-            // create package binding
-            assetHelper.withCleaner(wsk.pkg, packageName) {
-                (pkg, name) => pkg.bind("/whisk.system/alarms", name)
-            }
-
-            val futureDate = System.currentTimeMillis + (1000 * 30)
-            val payload = JsObject(
-                "test" -> JsString("alarmsTest")
-            )
-
-            // create trigger feed
-            println(s"Creating trigger: $triggerName")
-            assetHelper.withCleaner(wsk.trigger, triggerName, confirmDelete = false) {
-                (trigger, name) =>
-                    trigger.create(name, feed = Some(s"$packageName/once"), parameters = Map(
-                        "trigger_payload" -> payload,
-                        "date" -> futureDate.toJson,
-                        "deleteAfterFire" -> "true".toJson))
-            }
-
-            val actionName = s"$packageName/alarm"
-            val readRunResult = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "READ".toJson,
-                "authKey" -> wskProps.authKey.toJson
-            ))
-
-            withActivation(wsk.activation, readRunResult) {
-                activation => activation.response.success shouldBe true
-
-                    inside(activation.response.result) {
-                        case Some(result) =>
-                            val config = result.getFields("config").head.asInstanceOf[JsObject].fields
-
-                            config should contain("date" -> futureDate.toJson)
-                            config should contain("payload" -> payload)
-                            config should contain("deleteAfterFire" -> "true".toJson)
-                    }
-            }
-
-            val updatedFutureDate = System.currentTimeMillis + (1000 * 30)
-            val updatedPayload = JsObject(
-                "update_test" -> JsString("alarmsTest")
-            )
-
-            val updateRunAction = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "UPDATE".toJson,
-                "authKey" -> wskProps.authKey.toJson,
-                "trigger_payload" ->updatedPayload,
-                "date" -> updatedFutureDate.toJson,
-                "deleteAfterFire" -> "rules".toJson
-            ))
-
-            withActivation(wsk.activation, updateRunAction) {
-                activation => activation.response.success shouldBe true
-            }
-
-            val runResult = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "READ".toJson,
-                "authKey" -> wskProps.authKey.toJson
-            ))
-
-            withActivation(wsk.activation, runResult) {
-                activation => activation.response.success shouldBe true
-
-                    inside(activation.response.result) {
-                        case Some(result) =>
-                            val config = result.getFields("config").head.asInstanceOf[JsObject].fields
-
-                            config should contain("date" -> updatedFutureDate.toJson)
-                            config should contain("payload" -> updatedPayload)
-                            config should contain("deleteAfterFire" -> "rules".toJson)
-                    }
-            }
-    }
-
-    it should "update minutes parameter for interval feed" in withAssetCleaner(wskprops) {
-        (wp, assetHelper) =>
-            implicit val wskProps = wp
-            val triggerName = s"dummyAlarmsTrigger-${System.currentTimeMillis}"
-            val packageName = "dummyAlarmsPackage"
-
-            // the package alarms should be there
-            val packageGetResult = wsk.pkg.get("/whisk.system/alarms")
-            println("fetched package alarms")
-            packageGetResult.stdout should include("ok")
-
-            // create package binding
-            assetHelper.withCleaner(wsk.pkg, packageName) {
-                (pkg, name) => pkg.bind("/whisk.system/alarms", name)
-            }
-
-            val minutes = 1
-            val startDate = System.currentTimeMillis + (1000 * 30)
-            val stopDate = startDate + (1000 * 100)
-
-            // create trigger feed
-            println(s"Creating trigger: $triggerName")
-            assetHelper.withCleaner(wsk.trigger, triggerName) {
-                (trigger, name) =>
-                    trigger.create(name, feed = Some(s"$packageName/interval"), parameters = Map(
-                        "minutes" -> minutes.toJson,
-                        "startDate" -> startDate.toJson,
-                        "stopDate" -> stopDate.toJson))
-            }
-
-
-            val actionName = s"$packageName/alarm"
-            val readRunResult = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "READ".toJson,
-                "authKey" -> wskProps.authKey.toJson
-            ))
-
-            withActivation(wsk.activation, readRunResult) {
-                activation => activation.response.success shouldBe true
-
-                    inside(activation.response.result) {
-                        case Some(result) =>
-                            val config = result.getFields("config").head.asInstanceOf[JsObject].fields
-
-                            config should contain("minutes" -> minutes.toJson)
-                            config should contain("startDate" -> startDate.toJson)
-                            config should contain("stopDate" -> stopDate.toJson)
-                    }
-            }
-
-            val updatedMinutes = 2
-            val updatedStartDate = System.currentTimeMillis + (1000 * 30)
-            val updatedStopDate = updatedStartDate + (1000 * 100)
-
-            val updateRunAction = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "UPDATE".toJson,
-                "authKey" -> wskProps.authKey.toJson,
-                "minutes" -> updatedMinutes.toJson,
-                "startDate" -> updatedStartDate.toJson,
-                "stopDate" -> updatedStopDate.toJson
-            ))
-
-            withActivation(wsk.activation, updateRunAction) {
-                activation => activation.response.success shouldBe true
-            }
-
-            val runResult = wsk.action.invoke(actionName, parameters = Map(
-                "triggerName" -> triggerName.toJson,
-                "lifecycleEvent" -> "READ".toJson,
-                "authKey" -> wskProps.authKey.toJson
-            ))
-
-            withActivation(wsk.activation, runResult) {
-                activation => activation.response.success shouldBe true
-
-                    inside(activation.response.result) {
-                        case Some(result) =>
-                            val config = result.getFields("config").head.asInstanceOf[JsObject].fields
-
-                            config should contain("minutes" -> updatedMinutes.toJson)
-                            config should contain("startDate" -> updatedStartDate.toJson)
-                            config should contain("stopDate" -> updatedStopDate.toJson)
-                    }
-            }
-    }
 }
